check.cdf.type {affyio}	R Documentation

CDF file format function

Description

This function returns a text string giving the file format for the supplied filename

Usage

check.cdf.type(filename)

Arguments

filename
fullpath to a cdf file

Value

Returns a string which is currently one of:

text
the cdf file is of the text format

xda
the cdf file is of the binary format used in GCOS

unknown
the parser can not handle this format or does not recognize this file as a CDF file

Author(s)

B. M. Bolstad <bmb@bmbolstad.com>

[Package affyio version 1.56.0 Index]