read.cdffile.list {affyio}	R Documentation

Read CDF file into an R list

Description

This function reads the entire contents of a cdf file into an R list structure

Usage

read.cdffile.list(filename, cdf.path = getwd())

Arguments

filename
name of CDF file

cdf.path
path to cdf file

Details

Note that this function can be very memory intensive with large CDF files.

Value

returns a list structure. The exact contents may vary depending on the file format of the cdf file (see check.cdf.type)

Author(s)

B. M. Bolstad <bmb@bmbolstad.com>

[Package affyio version 1.56.0 Index]