read.celfile {affyio}	R Documentation

Read a CEL file into an R list

Description

This function reads the entire contents of a CEL file into an R list structure

Usage

read.celfile(filename,intensity.means.only=FALSE)

Arguments

filename
name of CEL file

intensity.means.only
If TRUE then read on only the MEAN section in INTENSITY

Details

The list has four main items. HEADER, INTENSITY, MASKS, OUTLIERS. Note that INTENSITY is a list of three vectors MEAN, STDEV, NPIXELS. HEADER is also a list. Both of MASKS and OUTLIERS are matrices.

Value

returns a list structure. The exact contents may vary depending on the file format of the CEL file

Author(s)

B. M. Bolstad <bmb@bmbolstad.com>

[Package affyio version 1.56.0 Index]