read.celfile.header {affyio}	R Documentation

Read header information from cel file

Description

This function reads some of the header information (appears before probe intensity data) from the supplied cel file.

Usage

read.celfile.header(filename,info=c("basic","full"),verbose=FALSE)

Arguments

filename
name of CEL file. May be fully pathed

info
A string. basic returns the dimensions of the chip and the name of the CDF file used when the CEL file was produced. full returns more information in greater detail.

verbose
a logical. When true the parsing routine prints more information, typically useful for debugging.

Value

A list data structure.

Author(s)

B. M. Bolstad <bmb@bmbolstad.com>

[Package affyio version 1.56.0 Index]