read.celfile.probeintensity.matrices {affyio}	R Documentation

Read PM or MM from CEL file into matrices

Description

This function reads PM, MM or both types of intensities into matrices. These matrices have all the probes for a probeset in adjacent rows

Usage

read.celfile.probeintensity.matrices(filenames, cdfInfo, rm.mask=FALSE, rm.outliers=FALSE, rm.extra=FALSE, verbose=FALSE, which= c("pm","mm","both"))

Arguments

filenames
a character vector of filenames

cdfInfo
a list with items giving PM and MM locations for desired probesets. In same structure as returned by make.cdf.package

rm.mask
a logical. Return these probes as NA if there are in the [MASK] section of the CEL file

rm.outliers
a logical. Return these probes as NA if there are in the [OUTLIERS] section of the CEL file

.

rm.extra
a logical. Return these probes as NA if there are in the [OUTLIERS] section of the CEL file

.

verbose
a logical. When true the parsing routine prints more information, typically useful for debugging.

which
a string specifing which probe type to return

Value

returns a list of matrix items. One matrix contains PM probe intensities, with probes in rows and arrays in columns

Author(s)

B. M. Bolstad <bmb@bmbolstad.com>

[Package affyio version 1.54.0 Index]